Manually creating service linked role
There are times when the role that creates the Lambda can't also have the IAM role to create the links. This means the role needs to be created manually. It took me a while to figure out the role name, I didn't realize I needed to parse it out from the ARN. Putting the CLI command in will make this quicker and easier.