Clarify DNS change vs CloudFront domain routing
When updating the DNS record to point to the new distribution, this will
not actually make the new distribution serve requests for that DNS
record. This is due to CloudFront's internal domain routing: When the
request reaches CloudFront, the request's Host header is handled by the
distribution that has a matching alternate domain name in its config.