Update lambda-examples.md
Found a small bug in the parseCookies() function. If the cookie contents is a base64 encoded string, and contains any trailing '=' signs, those are cut off by using the split function. This results in a cookie with "test=dGVzdDp0ZXN0MQ==" being parsed as `cookies['test'] = 'dGVzdDp0ZXN0MQ'` instead of `cookies['test'] = 'dGVzdDp0ZXN0MQ=='`

For Python this can be easily resolved by adding a limit 1 to the split function. Nodejs requires a different method, as doing a `cookie.split('=', 1)` results in the trailing '=' signs being removed from the string.