Highlighted distinction between caching of Status Codes and Custom error page
This distinction is very much required to understand and implement cache the error pages efficiently and thus the errors. Since the cloudfront still makes repeated calls to Origin even if it has cached the error pages. Also added the document link on this page for the solution to explicitly add the capability to modify Origin Responses which was missing.

We faced big issue due to error codes not being cached and cloudfront continuously hitting S3 and returning 4xx which broke the S3 request limit resulting in 5xx errors from S3. Original AWS Case ID 8556183081