Update the Cloudfront Function for routing
It's not uncommon or wrong to have directions with dots in their names.
This change updates the Cloudfront Function example to only restrict the last
path segment from not having any dots. It's still a bit annoying to having this
restriction but it's much better than restricting the full path from having a dot.